refactor updater
